+++ docs/scripts/util/crawlAllLinks.ts
@@ -18,6 +18,11 @@ async function checkSitemapPath(pageUrl: string, pageIdx: number) {
   let browser = await puppeteer.launch();
 
   const page = await browser.newPage();
+  // set a Chrome User-Agent header to make requests not appear from a headless browser
+  // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/User-Agent#chrome_ua_string
+  await page.setUserAgent(
+    'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36'
+  );
 
   await page.goto(pageUrl, { waitUntil: 'networkidle2' });
   await page.waitForNetworkIdle({ idleTime: 1000 });