+++ packages/e2e/cypress/integration/common/fileuploader.ts
@@ -37,3 +37,17 @@ When(
     );
   }
 );
+
+When('I drag and drop a file with file name {string}', (fileName: string) => {
+  cy.get('input[type=file]').selectFile(
+    {
+      contents: Cypress.Buffer.from('file contents'),
+      fileName,
+      lastModified: Date.now(),
+    },
+    /**
+     *  Since the input is hidden, this will need to be forced through Cypress
+     */
+    { force: true, action: 'drag-drop' }
+  );
+});