+++ packages/react-notifications/src/components/InAppMessaging/Backdrop/types.ts
@@ -1,0 +1,8 @@
+import { ViewProps } from '@aws-amplify/ui-react';
+
+export interface BackdropProps extends ViewProps {
+  /**
+   * Function called when backdrop is clicked
+   */
+  onClick?: React.MouseEventHandler<HTMLDivElement>;
+}